Fill:
-Etiqueta.java(entity)
-EtiquetaCtrl
-EtiquetaRepo(interface)
-Etiqueta service

According to business logic. Return "null"; must be substituted in a later commit for a valid html file

+++ src/main/java/com/cplcursos/java/kosso/entities/Etiqueta.java
@@ -1,6 +1,7 @@
 package com.cplcursos.java.kosso.entities;
 
 
+
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -18,9 +19,10 @@ public class Etiqueta {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Integer id;
+    Long id;
 
     private String nombre;
 
 }
+
